chore: remove `no-response` workflow
Having too many open issues is not a problem in this repository. It can easily be dealt with manually and as such be judged in a much more productive manner, avoiding needless closing / opening / closing as in eg #33

This PR essentially reverts #3

I would like to remove this from the rest of the eslint-community repositories as well, until we see a need.